requirements.txt
absl-py==0.9.0
astor==0.8.1
astunparse==1.6.3
beautifulsoup4==4.6.0
cachetools==4.1.0
certifi==2018.1.18
chardet==3.0.4
fire==0.3.1
gast==0.2.2
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
idna==2.6
importlib-metadata==1.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
lyricsgenius==1.8.6
Markdown==3.2.2
numpy==1.19.0
oauthlib==3.1.0
opt-einsum==3.2.1
pandas==1.0.5
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.6.0
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
regex==2017.4.5
requests==2.21.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
search-engine==0.0.6
six==1.15.0
soupsieve==2.0.1
tensorboard==1.15.0
tensorboard-plugin-wit==1.6.0.post3
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
toposort==1.5
tqdm==4.31.1
urllib3==1.22
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
